customer_back/requirements.txt
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
django-cors-headers==4.1.0